Add files via upload
This file analyzes simulated_data.csv, which has a similar structure to the WHI data example in the paper. The proposed estimator and standard errors are calculated, as are the naive estimator (based on self report alone), complete case biomarker regression calibration estimator, and the raking estimator which combines information from both self-report and biomarker.